# docs/swagger.yaml
basePath: /api/sf/v1
definitions:
  db.Pagination:
    properties:
      currentPage:
        type: integer
      size:
        type: integer
      skipPagination:
        type: boolean
      sort:
        type: string
      totalPages:
        type: integer
      totalRecords:
        type: integer
    type: object
...
  errs.ErrorType:
    enum:
    - OTHER
    - INVALID
    - IO
    - EXIST
    - NOT_EXIST
    - PRIVATE
    - INTERNAL
    - BROKEN_LINK
    - DATABASE
    - VALIDATION
    - UNANTICIPATED
    - INVALID_REQUEST
    - SMTP_ERROR
    - EMAIL_ERROR
    - STORAGE
    - INSUFFICIENT_ACCESS
    - CUSTOMER_DUPLICATE_TRANSACTION
    type: string
    x-enum-comments:
      BROKEN_LINK: Link target does not exist.
      CUSTOMER_DUPLICATE_TRANSACTION: Duplicate Transaction
      DATABASE: Error from database.
      EMAILERROR: SMTP Error
      EXIST: Item already exists.
      INSUFFICIENT_ACCESS: Insufficient Access
      INTERNAL: Internal error or inconsistency.
      INVALID: Invalid operation for this type of item.
      INVALID_REQUEST: Invalid Request
      IO: External I/O error such as network failure.
      NOT_EXIST: Item does not exist.
      OTHER: Unclassified error. This value is not printed in the error message.
      PRIVATE: Information withheld.
      SMTPERROR: SMTP Error
      STORAGE: Storage Error
      UNANTICIPATED: Unanticipated error.
      VALIDATION: Input validation error.
    x-enum-descriptions:
    - Unclassified error. This value is not printed in the error message.
    - Invalid operation for this type of item.
    - External I/O error such as network failure.
    - Item already exists.
    - Item does not exist.
    - Information withheld.
    - Internal error or inconsistency.
    - Link target does not exist.
    - Error from database.
    - Input validation error.
    - Unanticipated error.
    - Invalid Request
    - SMTP Error
    - SMTP Error
    - Storage Error
    - Insufficient Access
    - Duplicate Transaction
    x-enum-varnames:
    - OTHER
    - INVALID
    - IO
    - EXIST
    - NOT_EXIST
    - PRIVATE
    - INTERNAL
    - BROKEN_LINK
    - DATABASE
    - VALIDATION
    - UNANTICIPATED
    - INVALID_REQUEST
    - SMTPERROR
    - EMAILERROR
    - STORAGE
    - INSUFFICIENT_ACCESS
    - CUSTOMER_DUPLICATE_TRANSACTION
  errs.XError:
    properties:
      error: {}
      errorCode:
        type: string
      errorType:
        $ref: '#/definitions/errs.ErrorType'
      params:
        items: {}
        type: array
    type: object
...
  requestModel.Product:
    properties:
      categoryId:
        type: integer
      costPrice:
        type: number
      description:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      sellingPrice:
        type: number
      sku:
        type: string
    type: object
...
  response.DataResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/errs.XError'
      message:
        type: string
      pageMetaData:
        $ref: '#/definitions/db.Pagination'
      type:
        $ref: '#/definitions/response.ResponseType'
    type: object
  response.Response:
    properties:
      error:
        $ref: '#/definitions/errs.XError'
      message:
        type: string
      pageMetaData:
        $ref: '#/definitions/db.Pagination'
      type:
        $ref: '#/definitions/response.ResponseType'
    type: object
  response.ResponseType:
    enum:
    - Success
    - Failure
    - Warning
    type: string
    x-enum-varnames:
    - SUCCESS
    - FAILURE
    - WARNING
...
  responseModel.Category:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      productCount:
        description: Count of products in this category
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
...
  responseModel.Inventory:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      isLowStock:
        type: boolean
      lowStockThreshold:
        type: integer
      product:
        allOf:
        - $ref: '#/definitions/responseModel.Product'
        description: Related data
      productId:
        type: integer
      productName:
        type: string
      productSku:
        type: string
      quantity:
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
  responseModel.InventoryLog:
    properties:
      changeType:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      loggedAt:
        type: string
      loggedByName:
        description: User who logged
        type: string
      netChange:
        description: Calculated net change
        type: integer
      notes:
        type: string
      product:
        allOf:
        - $ref: '#/definitions/responseModel.Product'
        description: Related data
      productId:
        type: integer
      productName:
        type: string
      productSku:
        type: string
      quantity:
        type: integer
      reason:
        type: string
      stockAfter:
        description: Stock quantity after this movement
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
...
  responseModel.Product:
    properties:
      category:
        allOf:
        - $ref: '#/definitions/responseModel.Category'
        description: Related data
      categoryId:
        type: integer
      categoryName:
        description: Flattened category name
        type: string
      costPrice:
        type: number
      createdAt:
        type: string
      createdBy:
        type: string
      currentStock:
        description: From inventory
        type: integer
      description:
        type: string
      id:
        type: integer
      inventory:
        $ref: '#/definitions/responseModel.Inventory'
      isActive:
        type: boolean
      isLowStock:
        description: Stock alert flag
        type: boolean
      name:
        type: string
      sellingPrice:
        type: number
      sku:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
  responseModel.ProductAutoComplete:
    properties:
      currentStock:
        type: integer
      id:
        type: integer
      isLowStock:
        type: boolean
      name:
        type: string
      sku:
        type: string
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: This is the backend for Stitchfolio.
  title: Stitchfolio-backend API docs
  version: "1.0"
paths:
...
  /inventory-log:
    get:
      consumes:
      - application/json
      description: Get all inventory log records
      parameters:
      - description: search
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.InventoryLog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get all inventory logs
      tags:
      - InventoryLog
  /inventory-log/{id}:
    get:
      consumes:
      - application/json
      description: Get an instance of Inventory Log
      parameters:
      - description: Inventory Log id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.InventoryLog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get a specific Inventory Log
      tags:
      - InventoryLog
  /inventory-log/change-type:
    get:
      consumes:
      - application/json
      description: Get all inventory logs filtered by change type (IN, OUT, ADJUST)
      parameters:
      - description: Change Type
        enum:
        - IN
        - OUT
        - ADJUST
        in: query
        name: changeType
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.InventoryLog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get inventory logs by change type
      tags:
      - InventoryLog
  /inventory-log/date-range:
    get:
      consumes:
      - application/json
      description: Get all inventory logs within a date range
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.InventoryLog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get inventory logs by date range
      tags:
      - InventoryLog
  /inventory-log/product/{productId}:
    get:
      consumes:
      - application/json
      description: Get all inventory logs for a specific product
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.InventoryLog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get inventory logs by product ID
      tags:
      - InventoryLog
...
  /product:
    get:
      consumes:
      - application/json
      description: Get all active products with current stock
      parameters:
      - description: search
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get all active products
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Saves an instance of Product
      parameters:
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/requestModel.Product'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Save Product
      tags:
      - Product
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an instance of Product
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete Product
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Get an instance of Product with inventory
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get a specific Product
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Updates an instance of Product
      parameters:
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/requestModel.Product'
      - description: Product id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update Product
      tags:
      - Product
  /product/autocomplete:
    get:
      consumes:
      - application/json
      description: Autocomplete for products with stock info
      parameters:
      - description: search
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.ProductAutoComplete'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Autocomplete for products
      tags:
      - Product
  /product/low-stock:
    get:
      consumes:
      - application/json
      description: Get all products with stock below threshold
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get low stock products
      tags:
      - Product
  /product/sku:
    get:
      consumes:
      - application/json
      description: Get product details by SKU
      parameters:
      - description: Product SKU
        in: query
        name: sku
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseModel.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DataResponse'
      summary: Get product by SKU
      tags:
      - Product
...
swagger: "2.0"